pyproject.toml
[project]
name = "sports-commentator-agent"
version = "0.1.0"
description = "AI Sports Commentator - Vision Agents backend"
requires-python = ">=3.10"
dependencies = [
    "vision-agents[anthropic,openrouter,cartesia,deepgram,getstream]>=0.3.4",
    "python-dotenv>=1.1.0",
    "structlog>=23.0.0,<24",
    "rfdetr>=1.3.0",
    "supervision>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["agent/tests"]
